<REFACTOR> Phone을 PhoneNumber로 Rename
해당 서비스에서 저장하는 전화 관련 정보는 정확히는 전화번호이므로 PhoneNumber가 더 이해하기 쉬운 이름이라 판단되어 수정함
