+++ src/test/kotlin/com/samcho/user_authentication/domain/user/phone_number/SubscriberNumberTest.kt
@@ -1,4 +1,4 @@
-package com.samcho.user_authentication.domain.user.phone
+package com.samcho.user_authentication.domain.user.phone_number
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
